task_login_author_v1

+++ src/main/java/com/example/em/controller/ManagerController.java
@@ -2,6 +2,7 @@ package com.example.em.controller;
 
 import com.example.em.dto.manager.CreateManagerDTO;
 import com.example.em.dto.manager.ManagerDTO;
+import com.example.em.dto.response.ManagerLoginDTO;
 import com.example.em.dto.response.ResponseObject;
 import com.example.em.service.IManagerService;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -9,6 +10,8 @@ import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
+import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpSession;
 import java.util.List;
 
 @RestController
@@ -17,42 +20,69 @@ public class ManagerController {
     @Autowired
     private IManagerService service;
 
+    private Boolean isAuthor(HttpSession session){
+        boolean result = false;
+        ManagerLoginDTO manager = (ManagerLoginDTO) session.getAttribute("MA");
+        if(manager != null){
+            result = true;
+        }
+        return result;
+    }
 
     @GetMapping
-    public ResponseEntity<ResponseObject> getAll(){
-        List<ManagerDTO> result = service.getAll();
-        ResponseObject response = new ResponseObject(HttpStatus.OK.toString(), "List manager", result);
-        return ResponseEntity.status(HttpStatus.OK).body(response);
+    public ResponseEntity<ResponseObject> getAll(HttpSession session){
+        if(isAuthor(session)){
+            List<ManagerDTO> result = service.getAll();
+            ResponseObject response = new ResponseObject(HttpStatus.OK.toString(), "List manager", result);
+            return ResponseEntity.status(HttpStatus.OK).body(response);
+        }
+        ResponseObject response = new ResponseObject(HttpStatus.UNAUTHORIZED.toString(), "Vui lòng đăng nhập", null);
+        return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body(response);
+
     }
     @GetMapping("{id}")
-    public ResponseEntity<ResponseObject> getManagerById(@PathVariable Long id){
-        ManagerDTO result = service.getManagerById(id);
-        if(result != null){
-            ResponseObject response = new ResponseObject(HttpStatus.OK.toString(), "Manager details", result);
-            return ResponseEntity.status(HttpStatus.OK).body(response);
-        }else{
-            ResponseObject response = new ResponseObject(HttpStatus.NOT_FOUND.toString(), "Manager with ID " + id + " not found", null);
-            return ResponseEntity.status(HttpStatus.NOT_FOUND).body(response);
+    public ResponseEntity<ResponseObject> getManagerById(HttpSession session, @PathVariable Long id){
+        if(isAuthor(session)){
+            ManagerDTO result = service.getManagerById(id);
+            if(result != null){
+                ResponseObject response = new ResponseObject(HttpStatus.OK.toString(), "Manager details", result);
+                return ResponseEntity.status(HttpStatus.OK).body(response);
+            }else{
+                ResponseObject response = new ResponseObject(HttpStatus.NOT_FOUND.toString(), "Manager with ID " + id + " not found", null);
+                return ResponseEntity.status(HttpStatus.NOT_FOUND).body(response);
+            }
         }
+        ResponseObject response = new ResponseObject(HttpStatus.UNAUTHORIZED.toString(), "Vui lòng đăng nhập", null);
+        return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body(response);
+
     }
 
     @PostMapping
-    public ResponseEntity<ResponseObject> create(@RequestBody CreateManagerDTO managerDTO){
-        CreateManagerDTO result = service.addManager(managerDTO);
-        ResponseObject response = new ResponseObject(HttpStatus.CREATED.toString(), "Manager create successfully", result);
-        return ResponseEntity.status(HttpStatus.CREATED).body(response);
+    public ResponseEntity<ResponseObject> create(HttpSession session, @RequestBody CreateManagerDTO managerDTO){
+        if(isAuthor(session)){
+            CreateManagerDTO result = service.addManager(managerDTO);
+            ResponseObject response = new ResponseObject(HttpStatus.CREATED.toString(), "Manager create successfully", result);
+            return ResponseEntity.status(HttpStatus.CREATED).body(response);
+        }
+        ResponseObject response = new ResponseObject(HttpStatus.UNAUTHORIZED.toString(), "Vui lòng đăng nhập", null);
+        return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body(response);
     }
 
     @DeleteMapping("{id}")
-    public ResponseEntity<ResponseObject> deleteById(@PathVariable Long id) {
-        boolean isDeleted = service.deleteManagerById(id);
-        if (isDeleted) {
-            ResponseObject response = new ResponseObject(HttpStatus.OK.toString(), "Manager with ID " + id + " deleted successfully", null);
-            return ResponseEntity.status(HttpStatus.OK).body(response);
-        } else {
-            ResponseObject response = new ResponseObject(HttpStatus.NOT_FOUND.toString(), "Manager with ID " + id + " not found", null);
-            return ResponseEntity.status(HttpStatus.NOT_FOUND).body(response);
+    public ResponseEntity<ResponseObject> deleteById(HttpSession session, @PathVariable Long id) {
+        if(isAuthor(session)){
+            boolean isDeleted = service.deleteManagerById(id);
+            if (isDeleted) {
+                ResponseObject response = new ResponseObject(HttpStatus.OK.toString(), "Manager with ID " + id + " deleted successfully", null);
+                return ResponseEntity.status(HttpStatus.OK).body(response);
+            } else {
+                ResponseObject response = new ResponseObject(HttpStatus.NOT_FOUND.toString(), "Manager with ID " + id + " not found", null);
+                return ResponseEntity.status(HttpStatus.NOT_FOUND).body(response);
+            }
         }
+        ResponseObject response = new ResponseObject(HttpStatus.UNAUTHORIZED.toString(), "Vui lòng đăng nhập", null);
+        return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body(response);
+
     }
